requirements.txt
feedparser
feedgenerator
pytest
ruff
sqlalchemy
python-dateutil
flask
flask-cors
previewlink
APScheduler